Skips deleting deployments when acceptance tests fail
[#106774866]

+++ src/acceptance-tests/deploy/multiple_instance_rolling_deploy_test.go
@@ -38,8 +38,10 @@ var _ = Describe("Multiple instance rolling deploys", func() {
 	})
 
 	AfterEach(func() {
-		err := client.DeleteDeployment(manifest.Name)
-		Expect(err).NotTo(HaveOccurred())
+		if !CurrentGinkgoTestDescription().Failed {
+			err := client.DeleteDeployment(manifest.Name)
+			Expect(err).NotTo(HaveOccurred())
+		}
 	})
 
 	It("persists data throughout the rolling deploy", func() {
