+++ src/acceptance-tests/deploy/scale_down_instances_test.go
@@ -19,8 +19,10 @@ var _ = Describe("Scaling down instances", func() {
 	)
 
 	AfterEach(func() {
-		err := client.DeleteDeployment(manifest.Name)
-		Expect(err).NotTo(HaveOccurred())
+		if !CurrentGinkgoTestDescription().Failed {
+			err := client.DeleteDeployment(manifest.Name)
+			Expect(err).NotTo(HaveOccurred())
+		}
 	})
 
 	Describe("scaling from 3 nodes to 1", func() {
